requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
scikit-learn==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
arxiv==1.4.8
sentence-transformers
huggingface_hub